+++ src/context/SyntheticsStateContext/selectors/tradeboxSelectors/index.ts
@@ -1675,6 +1675,7 @@ export const selectTradeboxOffHoursLiqRisk = createSelector((q) => {
     marketInfo,
     isLong,
     nextSizeInUsd: nextPositionValues?.nextSizeUsd,
+    nextSizeInTokens: nextPositionValues?.nextSizeInTokens,
     nextCollateralUsd: nextPositionValues?.nextCollateralUsd,
     minCollateralUsd,
   });